perf(Rendering): Store colors etc as uchars instead of float
The original implementation only supported floats for
the VBO/CABO. This updates the code to use unsigned chars
by default for colors in the VBO. It also updates the stick
mapper to use unsigned chars for the offsets. The end result
is a 12 byte per point savings for color and an 8 byte per
point savings for offset.