fix(Rendering): some fixes and performance improvements
Clean up some of the rendering interaction code to better
manage animating and to ignore render calls when in the
middle of interacting. Also remove some calls that were
rendering the still image twice.

Also some code cleanup to improve performance some. Still a ways
to go in cases with hundreds of actors.