feat(Rendering): Add initial support for WebVR
Tested on Firefox with the Vive. Supports the notion
of WorldToPhysical coordinate systems and uses the
gamepad API to poll controller events. Currently the
right controller trackpad is used to control movement.

ResetCamera will compute reasonable values for PhysicalScale
and PhysicalTranslation.

Still needs to be tested against DayDream and other VR
systems. Support for their gamepads needs to be added.