fix(TubeFilter): Output scalar array size not matching expected
The output scalar array for the tube points was allocated using
FieldData::passData which adds the same size array as input points.
Other calls to passData with out of range indices would be silently
ignored. This change pre-allocates the output scalar array to the size
of the expected number of points in the output dataset.