fix(Rendering): skybox and cube map was borked
A number of fixes to cube maps and skyboxes. They are
slightly different in that one looks at a cube from outside
while the other looks at a cube form inside. This changes
how the texture coordinates are setup. Add to that the fact
that opengl uses an off negative Y image format for cubemaps
left ovber from Renderman standards.