fix(Rendering): add distortion pass, fix sizing, skybox
Add a second order radial distorion pass to vtk.js

To support this change the size queries on render
window to use the size of the active framebuffer
which may be the hardware buffer or a framebuffer
object.

Fix an issue in the skybox where the interpolaiton of
coordinates was incorrect. Most noticable for larger
view angles.