fix(Rendering): fix bugs in volume Rendering
Fix an issue where a texture is reused multiple times
with cropped data.

Fix an issue with the crop filter where the inputs
extent is not considered when computing offsets.

Fix a rendering issue where the mapping of the volume
to pixels was not quite correct resulting in deformations
when rendering cropped images.