feat(MouseCameraManipulator): Add new first person trackball manipulator
The MouseCameraTrackballFirstPersonManipulator allows a user to control the camera with the mouse in
the same way it is controlled in a first person shooter style game. Moving the mouse results in a
yaw and a pitch of the focal point. There is a pointer lock mode (enabled by default), which locks
the mouse after the render window has been clicked. The other mode requires left-click to be held on
the render window while moving the mouse.