feat(FirstPerson): Make the number of mouse movement animation steps configurable
Previously, the number of mouse movement animation steps was hard-coded to 10. Make this
configurable, so that users with slower computers can make it less (so it is not as laggy), and
users with faster computers can make it more (so it appears smoother). We have also changed the
default to 5.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>