feat(TextureLODsDownloader): Limit file size for textures
This provides a maxTextureLODSize option, that specifies the max KiB for textures. Any texture LODs
above this size will be skipped. The size checking is performed by making a HEAD request and
checking the content-length header. This seems to work well for chrome. However, for firefox, if the
request is cross origin (i.e., the origin of the html is different than the origin of the image),
the content-length header does not get returned for some reason, so it will not work for that case.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>