chore(ImageProperty): Rename scalarOpacity to piecewiseFunction
When used to render multi-component images slices using independent
components, the pwf is used as a component weighting function, rather than
to represent scalar opacity.  With dependent components, it is used as
scalar opacity.  Since the two cases do not overlap, we use one piecewise
function and texture for both.

The set/get scalarOpacity methods are maintained for backwards compatibility.