feat(RectangleWidget): add option to keep the representation plane orthogonal to the manipulator's n
The widget can have two points with different heights (usually when used with the imagemapper). This
can cause parts of the representation to be hidden when one of the two points is below the current
slice. This change adds an option to use the highest of the two points to set the representation
plane's height, to make sure that the representation stays on top of the ImageMapper.