feat(custompostprocessingpass): allows 2D convolution kernels to be applied to the rendered image
The class vtkCustomPostProcessingPass was added. The class vtkCustomPostProcessingPass can be
integrated in into the vtk.js render-pipeline to enable 2D image enhancements on the rendered
images. This works by setting the vtkRenderPass as a delegate pass and then using the
vtkCustomPostProcessingPass as the render pass on the vtkRenderWindow. Any valid
2D-convolution-kernel can be used.