feat(convolution2dpass): allows 2D convolution kernels
The class vtkConvolution2DPass was added. The class vtkConvolution2DPass can be integrated in into
the vtk.js render-pipeline to enable 2D image enhancements on the rendered images. This works by
setting the vtkRenderPass as a delegate pass and then using the vtkConvolution2DPass as the render
pass on the vtkRenderWindow. Any valid 2D-convolution-kernel can be used.