feat(Convolution): Allows 2D convolution kernels (#1549)
Allow 2D convolution kernels to be applied to the rendered image

The class vtkConvolution2DPass was added. The class vtkConvolution2DPass can be integrated into
the vtk.js render-pipeline to enable 2D image enhancements on the rendered images. This works by
setting the vtkRenderPass as a delegate pass and then using the vtkConvolution2DPass as the render
pass on the vtkRenderWindow. Any valid 2D-convolution-kernel can be used.

Co-authored-by: Dominik Werner <dwerner@casra.ch>