chore(WebGPU): add a UBO class and use it
Add a Uniform Buffer Object class for WebGPU to manage UBOs
and make it easier for people to extend mappers etc by adding
additional fields to the UBOs used in a pipeline.

The UBO class handles memory alignment issues but currently
only supports basic types such as vec/mat not arrays or
structs as members.