chore(WebGPU): remove use of window
Remove the use of window to create native arrays throughout vtk.js. Also
check for cases where a dataArray is created with a type of Array which
is invalid and promote to Float64Array instead.

Make the use of Map in models to use _ as a prefix
to indicate it is protected and not to be serailized.