feat(WebGPU): add hardware selector support
add supporrt for actor/composoite picking in WebGPU

As WebGPU is a more asynchronous API this required creating
a new API for selections that is async. In this case the
way to do selections is

const srcData = await publicAPI.getSourceDataAsync(fx1, fy1, fx2, fy2);
if (srcData) {
  return srcData.generateSelection(fx1, fy1, fx2, fy2);
}

In the call to getSourceData you can get the entire window and
then make multiple calls to generate selection on the result.

Note that generateSelection is called on the result, not on the
selector as there may be multiple selections in flight at a time
so the state of the selector may not be correct for a specific
selection. All required state for generateSelection is kept in
the retuened srcData object.

Reworked the WebGL hardware selector to support the new API
in addition to the old API.

Rewrote the hardware selector example to use the new API.

Left the other couple cases of using ther hardware selector
using the old API for now as making them async could be a
breaking change.