feat(WebGPU): add support for testing
Fixed issues and added a working CaptureNextImage so
that the standard vtk.js testing structure can be used
and works. A number of tests are failing but it looks
like the expected tests are failing.

Fixed viewport support so that multiple renderes can now
work together properly. This involved creating an opaque
pass and adding a clear method that can be used to clear
the viewport instead of relying on the loadValue of the
render encoder (which gets called too often anyhow) the
multiple renderer test now passes on WebGPU.

Cleaned up one aspect of how UBOs are used. In a couple
places the group of the UBO was hard coded, now the group
for the UBO is always queried from the pipeline. This is
a safer and more general approach.

Added support for the FullScreenQuad class to have a UBO.

Updated many tests that explicitly required webGL to now
be generic. Exclude them from the test:webgpu script.