fix(webgpu): fix some issues and add hardware select tests
Move hardware selector tests to Rendering\Core and update
them for the new async selection API

Fix a bug where UBOs were always being considered modified
when doubles were converted to different floats

Fix an issue where adding the fragment based clear approach
broke hardware selection

Add support for compositeID on the hardware selector and
implement it for the glyph mapper.

Fix an issue with the WebGPU hardware selector where it was
incorrecly addressing the zbuffer (needed to invert y)

Fix an issue where the vertex input could be reordered causing
the vertex shader to get mismatched inputs. Not the vertex
input is always sorted, so that the order arrays are added
is irrelevant.