fix(httpdatasetreader): when caching array data, differentiate by class
The reader caches array instances to avoid having to load the same array multiple times. Arrays are
cached using a hash of the binary data they contain. This means a vtkPoints and vtkDataArray sharing
the same data could be cached under the same index. This case can happen quite often with empty
meshes, as all arrays will share the same hash. This appends the class type to the caching key to
avoid these conflicts.