chore(WebGPU): work on multi volume rendering
Add basic structure for multi volume rendering

Rework a lot of the structure of the WebGPU support.

1) Make BindGroupLayouts more generic where you request
them with a description and the device caches them and
either returns or creates as needed.

2) make bindable things more generic so they can
be lumped into one group and treated the same. This
include UBOs, SSBOs, and TextureViews

3) add a mapper helper that is meant to be a minimal mapper.
The idea is that PolyDataMapper would internally use four of
these simpler mappers, one for each cellArray (verts, lines,
polys, strips). Then other classes could make use of the simpler
mapper helper directly when such complexity isn't needed.

4) Make the renderencoder the more important part of rendering.
Previously some calls were passed the pipleine that now get
passed the render encoder. The renderencoder has a bound
pipeline so while those two things (pipeline and renderencoder)
are a bit independent of each other, during rendering the
renderencoder will have a pipeline and tends to live (be bound)
longer than the pipeine so it is what gets passed.