chore(WebGPU): more volume support, add bind group
Flesh out the multivolume renderer a bit more

Add a bind group class as WebGPU only requires support for 4 bind group
layouts and the old approach could easily go past 4. With the bind group
class we typically only have two bind groups active, one for the
renderer and one for the mapper. Tons of changes to support this.