fix(macro): Apply macro.js shim at pkg build time
babel-plugin-macros interferes with our macro.js file. As it is not
ideal to rename it, nor is it feasible to convince babel-plugin-macros
to change its "ownership" of any files ending in "macro.js", this is an
attempt at shimming macro.js to effectively sneak past
babel-plugin-macros while still maintaining compatibility with existing
macro.js consumers.