fix(WebGL): fix handling of ipScalarRange with halfFloat
ipScalarRange is a 0-1 value in texture values which with
different texture formats may require different scaling/offsets.

This commit fixes it's use for halfFloat textures and also
changes how it is implemented to be done with uniforms
which are quick to change, instead of changing the shader program
which is expensive.