fix(rendering): Correction to domain boundary for images and volumes
(#1936)
Previously, in a vtk representation of an image/volume, pixels or voxels
are considered to exist on the points of a regular grid. This results in
an error of half-pixel/voxel length in the bounds of the rendered object
(plane or box that acts as a domain of the image/volume). While openGL
texture mapping automatically adds the half-voxel offset, we  have
corrected the size of the rendered domain to correctly represent the
bounds of the image data.