fix(Picking): fix a picking bug with WebGL and WebGPU
An argument was missing from a call in the vtkPicker pick method

Views were modifying the data (renderables) by performing
camera creation and reset. Views should not modify the data
(renderables). This was also causing a picking failure for WebGPU
as one render was not enough to invoke the camera reset code.

Updated some tests that were calling traverseAllPasses instead of
render.  We want to limit interaction with API specific windows.

Invoke preRender in the hardware selectors for now since they
bypass render and directly call traverseAllPasses. In the future
we should change this

https://github.com/Kitware/vtk-js/issues/2122