feat(WebGPU): use invereted depth and improve edges
Switch the zbuffer on WebGPU to be an inverted
depth buffer. Update mappers to handle this. See the README
for more information on how it works.

Improve surface with edges by using thicker edges and
a tiny bit of offset.

Update for recent changes in WebGPU that require flat interpolation
for integral outputs/inputs.