feat(labelwidget): add new-style LabelWidget
vtkLabelWidget was one of the last widgets to exists only in the old-style/deprecated version
(`Interactions/Widgets`). This commit implements a new vtkLabelWidget, alongside with a new example.
It also add a new property `strokeFontProperties` to SVGLandmarkRepresentation and fix its method
`fontSizeToPixels`.
In SVGLandmarkRepresentation, `textProps.fill` property was replaced by `fontProperties.fontColor`
(to be more consistant with the other properties), and textProps was removed from default values.