fix(macro): enforce the creation of a copy when array is passed to set accessor
When calling obj.setMyProp(myArray), a copy of the array was not made
(a reference of the same array was kept). Changing myArray was therefore
also modifying the obj model.myProp.

Note that it breaks reslice cursor when camera viewUp was set outside
of the widget (e.g. with ctrl+left mouse)