refactor(WebGL): remove support for WebGL1
Now that safari supports webgl2 move to that as the webg1 code
and conditionals added a lot of complexity and limited what
we could do.

Now that WebGL2 is fairly universal this change should not
break anything other than iOS devices still running iOS 14
or earlier.

There is a second step (not in this MR) which is to update vtk.js
to assume that we have depth textures. This will remove the need
for the opaque depth pass code and some other sections of code
but to keep this MR reasonably small it is being left out.