feat(WebGPU): better support for multiple volumes
This MR adds more gracefull handling for the case when
there are many volumes. Now any number of volumes can be rendered
and up to a hardware dependent number of them can be correctly
intermixed. Typically 12 volumes can be intermixed. When more than
12 volumes are present they are rendered in 12 volume groups
from back to front.

Fixes an issue in the volume FSQ when the number of volumes
was changing.

Related to https://github.com/Kitware/vtk-js/issues/2086