feat(WebGPU): support interactive volume rendering
The volume pass now uses the recentAnimaitonFrameRate
to render to a smaller texture when the view is animating to try
to reach the DesiredUpdateRate.

Added a new approach for checking frame rate to the
RenderWindowInteractor as web browsers are becoming more
strict about reporting exact elapsed times (due to fingerprinting)
so the new appraoch is designed to use a longer time frame thus
avoiding the anti fingerprint rounding the Date.now() does.