perf(WebGL): rework WebGL volumeMapper interactive rendering
Update the renderwindowinteractor to set the recentFrameTime
from the new approach which uses at least a one second
timeframe to compute rendering rates.

Update the WebGL VolumeMapper to use a callback for when the
animation rate changes on the renderwindowinteractor
and simplify a bit how the use of the smaller viewport
is handled.

Adjuts the volumeviewer example to be 30 fps because 15fps
feels slow.