fix(pick): null actor when returning selection
In `vtkOpenglHardwareSelector`, functions `getPixelInformation()` and
`getPixelInformationWithData()` return a
valid flag when actor id read from the pixel buffer is positive.
However, the actor id may not correspond to a valid actor object stored
in the model.props array. An additional check is added for validating
that the actor id corresponds to a valid array index. Else, a null
selection is returned. This fix is related to a corresponding fix in
react-vtk-js [PR 47](https://github.com/Kitware/react-vtk-js/pull/47)
for picking the closest object from the mouse pointer position.