fix(Rendering): Fix ScalarBarActor text generation
The old code used an async method to get the text image
from the canvas for the letters. The new code uses an
approach that doesn't require an asynchronous call.

This adds support for creating textures from a canvas
or a JS ImageData object in addition to vtkImageData and
a JS Image as before.