fix(getstate): fix getState() on deleted object and null content
Calling getState() on a deleted object was crashing because
model.classHierarchy was never considered as null.
It can be null because Karma calls serialize() (and now getState() due to the toJSON() addition)
on all test results (such as t.ok(instance) in
test 'Test vtkAppendPolyData instance') after they are deleted.

getState() was also crashing on vtk objects that contained null child objects
such as filters with no input connection or input data,
e.g. with vtkAppendPolyData: Object{
  classHierarchy: ['vtkObject', 'vtkAppendPolyData'],
  inputArrayToProcess: [],
  inputConnection: [null],
  inputData: [null],
  numberOfInputs: 1,
  output: [],
  outputPointsPrecision: 0,
  vtkClass: 'vtkAppendPolyData'
}

getState() was also crashing on ProxyManager.