feat(imagemapper): ability to use lookup table scalar range
This change adds functionality to map pixel data to colors using one of two mechanisms in the image
mapper:

- Use color window/level values from the image property
- Use the range set on the lookup table in the image property

By default, the color window/level values are used.

BREAKING CHANGE: The original behavior of the image mapper was that if a lookup table is provided,
it mapped the lookup table's scalar range by default. The new behavior disables using the lookup
table scalar range by default. Instead, the window/level values are used.