chore(WebGPU): cleanup webgpu mappers
Cleanup and document WebGPU mapper architecture.
This is a first step in trying to make the mappers
for WebGPU cleaner, better documented and more extensible.
The main trust here is to create a CellArrayMapper and update
the polydatamapper to just instantiate and use that. For example
one cellarray mapper for each of verts, lines, polysm, and strips
as opposed to having the polydatamapper try to manage all that
internally.

Also renamed and documented many methods to be more clear, fit better
with WebGPU terminology and have a more clear defined process.

There is more work to be done, the goal being to make mappers both
1) easy to customize - add shader code, add arrays to the VertexInput
2) easy to extend - with data pipelines we can easly add more stages
(filters) to data processing in VTK in a chain but with mappers more
capabilities means a larger and larger bloated mapper, the end goal
would be to enable mappers to be extended without adding to their
base code, but more like adding a filter chain.

Update to latest WebGPU standard