refactor(math): number[][] to number[] (matrices)
change matrices format from number[][] to number[] in math module and
where the modified functions are called. Tests to the functions modified
have been added

BREAKING CHANGE: For functions calling math functions for computations
with matrices, the format must now be number[] or Matrix as defined in
the typescript definitions.