feat(manipulator): harmonize and improve widget manipulators
Harmonizes the way manipulators are used in widgets so its more consistent. Improves the
manipulators so they can use 3 different origins: `userOrigin` (set by the user outside of the
widget code), `handleOrigin` (automatically set if the handle have a manipulator mixin) and
`widgetOrigin` (occasionally set in the widget code to give a "global" origin to the manipulator).
The manipulator normal is handled the same way, but can also use camera normal if the property
`useCameraNormal` is set to true.