fix(web-worker-loader): patch/fix vite prod build
rollup-plugin-web-worker-loader funcToSource does not work with minified
functions, which gets it into trouble when a minifier does a pass over
the webworker code. This patch is intended to work around this issue by
wrapping the function in a function call, since the results of
funcToSource are evaluated in a worker context in the end.