fix(WebGPU): fix to caching and minor fix to stick mapper
Stick mapper was setting frag_depth inside an if statement and
that was not working. Moved it outside the if statement.

Updated the device cache to use a map of weakrefs instead
of a weakmap. The old approach had issues keeping objects
too long if their sources never were freed. Such as if an
array just changes mtime but not source (not the best approach
but it happens)  The new cache approach just uses a WeakRef to
avoid the issue with sources/owner and so that concept was removed.

Now it is just a hash -> weakRef(object) map approach and when the
object finalizes we remove the entry in the map.