feat(spline1D,spline3D): open splines
Two methods implemented (Kochanek, Cardinal). Both methods need boundary conditions. Different
conditions are available (on the derivative, the second derivative or the second derivative on the
interior point). Values for those conditions can be chosen. The boundary conditions as well as the
value for it are the same on both sides but could be separated in further implementations.  The
spline widget and its example have been updated.