refactor(fullscreenrenderwindow, dataarray, remoteview, macro): refactor constructors
Refactor of (some classes) constructors so that the setXXX() methods of coresponding properties are
called when the classes are instanciated (calling newInstance). In the macro.js newInstance()
function, the set() method is called in order that the setXXX() methods for each entry of
initialValues is called if such methods exist. This is in order that side effects of setters are
taken into account when creating new objects (for example through the call of modified()). Setters
of both default and initial values are called.
Refactor of DataArray and child classes. The method "setData" is called so that the parameters are
synchronised as it is not a setter of one property in itself but has that purpose. Warning : it
cannot be overwritten in child classes without possible issues of behavior when creating the object.
Tests have been added and updated.
A new test for macro protected properties have been added.

BREAKING CHANGE: If a custom handling is done on the initialValues in the extend function, the
changes made might be overwritten by the call of the set function in newInstace. If it is not what
is intended, those properties have to be removed from initialValues after custom processing. If a
setter of a property is called in the function creating an object (vtkObject(publicAPI, model)) such
call should be removed to avoid witnessing twice the same side effects.