fix(WebGPU): fix two issues in WebGPU
Fix an issue where the renderer UBO was being shared between
pipelines causing some multiviewport renderings to use an
incorrect UBO

Fix an issue in the ImageMapper where the world coordinates
and texture coordinates were not consistent with the new
definition of pixel ceneterd images.

Fix an issue where the range of a color transfer function
was being ignored in one place while used in building the table.