refactor(mousewheel): normalize mouse wheel event spin values
This attempts to fix the problem of not being able to scroll a volume by 1 slice for each mouse
wheel.
Depending on the OS and the browser, doing a mouse wheel event typically
procuces a spin value > 1 (always <1 for trackpad).
This can be explained because mouse wheel events are designed to "scroll
a page by multiple lines at once".
In VTK.js specific domain, having a scroll spin value of 1 makes sense
by default, hence the normalization factor.