fix(MouseRangeManipulator): asymmetric scroll wheel behaviour
Fix a bug where the scrollListener behaves in an asymmetric manner
when comparing increasing value against decreasing value.
This occured due to the use of Math.round() function
and its application to the independently calculated newValue.
Additionally, Math.round() also behaves asymmetrically when applied
to the value of 0.5, as it always rounds towards +infinity.

The correct approach is to calculate the number of accumulated steps
from the newDelta and round down to the closest lower integer
using Math.floor().
This new approach also requires that the listener.step value be
always positive. A required check has been added to the input value.
If the caller needs to flip the scroll wheel direction, they
should use negative scale values instead.

Since, we are now using the scale as signed value,
add a line to preserve the sign of scale after the power operation
for exponential scaling.