fix(RemoteMouse): send position in wheel event
This modifies the remote mouse events to use `createRemoteEvent()` similarly to the other mouse
events. One thing this adds is the `x` and `y` positions of the mouse, which can be used to
determine which renderer the mouse is hovering over.

This is needed as a first step to fix pyvista/pyvista#4020

The next step is to fix the web protocol to use the mouse position to determine which renderer
the mouse is over.

I have tested this locally and it appears to work properly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>