fix(lights): do not add duplicate lights
In trame-vtk, we have a call to add a light.

Sometimes, this would be triggered multiple times, like if the user toggled
local view on/off.

It doesn't really make sense to add the same light twice (you can just modify
the intensity instead). So, similar to the view prop check,
add a check to prevent the same light from being added twice.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>