fix(WidgetManager): Support multiple renderers in the same render window
To reproduce the bug, 2 renderers must belong to the same renderwindow
and a widget must be added to both renderers.
When mousing over the first renderer, the widget manager was making
the widget active, but the widget manager of the second renderer was
deactivating the widget because the mouse was not over the widget in the
second renderer.