perf(WebXR): Unify WebXR scene default physical factors
Updates default WebXR scene factors to standardize across
VR and AR experiences. Factors are selected to scale and
translate the default scene so that VR and AR example datasets
can be viewed comfortably. Previous behavior resulted in
VR models appearing very large and very close to the user's
face, which was particularly uncomfortable in testing on the
HoloLens 2 headset. In updated behavior the examples in
`GeometryViewer` and `WebXRVolume` appear smaller and at a
comfortable distance from the user.