test(ImageResliceMapperShareOpenGLTexture): fix broken test
Fix the broken test due to changes to the slicing-plane polydata
construction in vtkImageResliceMapper.
Remove outline from the test as it is not essential for the test,
and the outline filter requires fix to use ImageData transform
rather than simply using the axis-aligned dataset bounds.

Compensate for scaling of VBO vertices when doing computations
with normals.