fix(getstate): serialize typed arrays as array
the behavior of JSON.stringify for typed arrays is to serialize them as objects with numeric keys.
This cause crashes when re-building the object since it would try to fit an object where a typed
array is required. This changes makes it so getState replaces typed arrays with plain javascript
arrays.