fix(reslicecursorwidget): fix camera reset on rotation after a camera translation
When the user translates a camera (shift+left drag), and then rotate the reslice cursor widget in
another view, the camera was reset. This is because the focal point offsets were not "stored" after
the camera translation. The rotation was therefore resetting the camera translation. A solution
would have been to observe the interactor style each time the camera is manipulated. Instead, we
prefer to save the focal point offsets before any reslice cursor widget interaction happens, namely
when startInteraction is triggered. This makes computeFocalPointOffset parameter in the reslice
cursor widget interactionEvent obsolete. It has therefore been removed along with the associated
canUpdateFocalPoint. This is not the behavior's responsibility to tell the event consumers what they
must do.

BREAKING CHANGE: ResliceCursorWidget.interactionEvent has been modified