feat(PolyDataMapper): Use a texture and texture coordinates for cell colors
When coloring using cell data, colors per point were pre computed and sent to the VBO.
They now use the same mechanism as for point data:
texture coordinates in the transfer function are sent in the VBO and the colors are in a texture.
This means that if only the colors change, but not the range of the LUT, the VBO is not recreated.