fix(picking): write to depth buffer for translucent pass when picking
This change fixes a picking issue with multiple translucent actors overlaid on top of each other.
The hardware selector would just register the propID of the last rendered prop. This was because we
explicitly turn off writing to the depth buffer when rendering translucent props as OIT doesn't need
it. This change enables writing to the depth buffer when picking in the translucent pass.

BREAKING CHANGE: Applications relying on picked propID would see a difference in the IDs returned by
the hardware selector.

fix #2800