fix(ImageResliceMapper): Fix offset of WCTCMatrix
The `indexToWorld` matrix transforms an integer index coordinate into the center of a voxel.
The WCTCMatrix transforms world coordinates to texture coordinates.
Therefore we need to add  a 0.5 voxel offset (see `getSpatialExtent`) to this matrix.
We need this offset because the voxel region is all around the voxel coordinate instead of after it.
This is similar to what is done in ITK, see this link:
https://itk.org/ITKSoftwareGuide/html/Book1/ITKSoftwareGuide-Book1ch4.html#x45-540004.1.4

Co-Authored-by: yarous224