feat(cellDataColors): Use 3D textures when needed for cell data colors
When coloring by point data, with interpolateScalarsBeforeMapping set to true, we can't use a
3D texture because we interpolate texture coordinates.

When coloring by cell data, there will be no texture coordinate interpolation as all the points of
a cell have the same texture coordinates. Therefore we can use 3D texture to be able to use more
values. We need to be able to interpolate between two neighboring texels, this is why we have to
use a zigzag pattern for the color texture.

The theoretical limit for the number of voxels is 2048**3 = 2**33.
We could also use the 3D texture when not interpolateScalarsBeforeMapping is false, but this is
not done in this PR.

BREAKING CHANGE: Removed `mapScalarsToTexture` from the public API.