feat(colortransferfunction, mapper): logarithmic scales support for ColorTransferFunction
Allows log scaling for ScalarBar as well as as Mappers using ColorTransferFunction.

BREAKING CHANGE: Previous log scaling was disabled but functionality is now quite different from
what it maybe was before it was disabled. ColorTransfer function now maps the model.nodes in
logarithmic scales so there's an even spacing between powers of ten.