fix(OpenGLTexture): fix regression causing RGB DICOM not to display
A regression was introduced during the transition to named parameters,
where numComps was passed instead of numberOfComponents set to 1
by default. This caused RGB DICOM images to display incorrectly, as only
one component was used instead of the intended three.

This change restores the correct handling of numberOfComponents,
ensuring that multi-component (RGB) textures are displayed properly.

Ref: https://github.com/kitware/vtk-js/commit/7fbbaa83b697f7fb7b09fe03a05fbca397271496

Fixes #3344