fix(Glyph3DMapper): use correct shift-scale transform
Previous function to apply shift and scale transform to the
existing glyph matrix was incorrect
because it did not account for scaling of rotation components.
The glyph matrices can contain rotation components, which
when combined with a shift-scale operation will get affected by
scaling.

It is therefore, better to use proper matrix multiplications
rather than trying to recreate the effect through manipulation of
individual components of a matrix.